TB-305 WIP:  add delete existing apps
