+++ packages/portal-admin/package.json
@@ -59,6 +59,7 @@
     "@types/webpack-env": "^1.16.2",
     "axios": "^1.6.2",
     "dotenv": "^16.3.1",
+    "notistack": "^3.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "single-spa": "^5.9.3",
